renaming business plan to richard to fit with other naming conventions
